Types

This page shows all the types defined in Sym4state.Types.

Module

Sym4state.Types — Module

This module defines structures containing crystal structures, symmetric operations as well as relation among different magnetic configurations.

This module exports:

CoeffMatRef

Map

Struc

SymOp

Types and methods

Sym4state.Types.Atom — Type

The data structure designed to store an atom includes its atomic number, position, and spin.

Two Atoms can be deemed approximately equal using the isapprox function if they share the same atomic number, possess position vectors that are approximately equal, and have spin vectors that are approximately equal.

struct Atom{T<:AbstractFloat}

num::Int64: Atomic number.

pos::StaticArraysCore.SVector{3, T} where T<:AbstractFloat: Position vector in fractional coordinates.

spin::StaticArraysCore.SVector{3, T} where T<:AbstractFloat: Spin direction vector in cartesian coordinates.

Sym4state.Types.Struc — Type

The data structure created for storing a material structure.

struct Struc

uni_num::Int64: Unique index number to identify the material.

lattice_mat::StaticArraysCore.SMatrix: The lattice matrix [a b c].

atom_count::Int64: Number of atoms in this structure.

num_vec::StaticArraysCore.SVector: Atomic numbers for different atoms.

pos_mat::StaticArraysCore.SMatrix: Fractional coordinates for different atoms.

spin_mat::StaticArraysCore.SMatrix: Spin directions for different atoms.

Sym4state.Types.Struc — Method

A convenient approach to construct a new Struc with essential parameters exclusively:

Struc(
    uni_num::Int64,
    lattice_mat::AbstractArray{T, 2},
    num_vec::AbstractVector{Int64},
    pos_mat::AbstractArray{T, 2},
    spin_mat::AbstractArray{T, 2}
) -> Sym4state.Types.Struc

Sym4state.Types.SymOp — Type

The data structure created for storing a symmetric operation.

struct SymOp{T<:AbstractFloat}

rot_mat::StaticArraysCore.SMatrix{3, 3, T} where T<:AbstractFloat: The rotational matrix for atoms.

spin_rot_mat::StaticArraysCore.SMatrix{3, 3, T} where T<:AbstractFloat: The rotational matrix for spins.

trans_vec::StaticArraysCore.SVector{3, T} where T<:AbstractFloat: The translation vector for atoms.

proper::Bool: The indicator flag to denote the propriety of the rotation.

time_rev::Bool: The indicator flag to denote the presence of a time reversal operation.

Sym4state.Types.SymOp — Method

A convenient approach to construct a new SymOp with essential parameters exclusively:

SymOp(
    rot_mat::AbstractArray{T, 2},
    trans_vec::AbstractArray{T, 1},
    time_rev::Bool,
    lattice_mat::AbstractArray{T, 2},
    inv_lattice_mat::AbstractArray{T, 2}
) -> Sym4state.Types.SymOp

As the spins are pseudovectors, their rotational matrix can be defined as:

\[\mathbf{R}^\prime = T \cdot \|\mathbf{R}\| \cdot \mathbf{R}\]

where $\mathbf{R}$ is the rotational matrix for atoms and $T$ is the time reversal operation.

Sym4state.Types.Map — Type

The data structure created for storing all the configurations needed to calculate an interaction matrix.

struct Map

map_mat::Matrix{Vector{Int8}}: A matrix encompassing all the uni_num values of the configurations required to compute the elements of the interaction matrix.

fallback_vec::Vector{Int8}: A vector that stores the connections between the indices of the required real configurations within the set of all configurations.

struc_vec::Vector{Sym4state.Types.Struc}: A vector that stores all the magnetic configurations.

type::Int64: The type of this interaction matrix. 1 for exchange interaction matrix while 2 for single-ion anisotropy matrix.

Sym4state.Types.Map — Method

A convenient approach to construct a new Map with essential parameters exclusively:

Map(
    fallback_ds::DataStructures.IntDisjointSets,
    all_struc_vec::Vector{Sym4state.Types.Struc};
    rotation_symmetry_flag
) -> Sym4state.Types.Map

Sym4state.Types.CoeffMatRef — Type

The data structure created for storing the equivalent relation between two atom pairs.

struct CoeffMatRef

group_idx::Int64: Unique index number to identify the group of those two atom pairs.

pair_vec::AbstractVector{Int64}: A vector indicating two atom pairs.

op::Sym4state.Types.SymOp: The operation linking those two atom pairs.