Utilities

This page shows all the functions defined in Sym4state.Utils.

Module

Sym4state.Utils — Module

This module defines useful functions for simplifying the four-state method.

This module exports:

check_unit_cell

check_z_rot

equal_pair

get_all_interact_struc_vec

get_fixed_pair_vec

get_pair_and_coeff

get_py_struc

get_sym_op_vec

linear_idx_to_vec

magonly

py_struc_to_struc

struc_compare

to_vasp_inputs

Functions

Structures and configurations

This section contains functions working with structures or magnetic configurations.

Sym4state.Utils.mag_config — Function

This function yields an array with size of [3, mag_count, 36] that comprises all 36 magnetic configurations utilized in the computation of the interaction matrix for a given number of magnetic atoms mag_count and the initial norm value s_value of MAGMOM.

The resulting configurations pertain to the interaction between two atoms specified by the target_idx_vec vector.

mag_config(
    mag_count::Int64,
    target_idx_vec::Vector{Int64},
    s_value::Float64
) -> Array{Float64, 3}

Sym4state.Utils.py_struc_to_struc — Function

This function transforms the Structure object from Pymatgen into a Struc object.

The distinct identifier uni_num will be supplied to the Struc object. The spin_mat variable represents the spin value assigned to each atom in the structure.

py_struc_to_struc(
    py_struc;
    uni_num,
    spin_mat
) -> Sym4state.Types.Struc

Sym4state.Utils.get_all_interact_struc_vec — Function

By utilizing the mag_config, this function generates all the Struc objects required for calculating the interaction matrix.

It takes the atomic numbers of magnetic elements stored in mag_num_vec, the index of target atoms denoted by target_idx_vec, and the initial value s_value.

get_all_interact_struc_vec(
    struc::Sym4state.Types.Struc,
    mag_num_vec::Vector{Int64},
    target_idx_vec::Vector{Int64},
    s_value::Float64
) -> Vector{Sym4state.Types.Struc}

Sym4state.Utils.struc_compare — Function

This function examines two Struc objects and determines their approximate equivalence within a tolerance of atol.

It returns a Tuple that includes a boolean value indicating whether they are equal, as well as the corresponding relationships among the atoms in the respective structures.

struc_compare(
    x::Sym4state.Types.Struc,
    y::Sym4state.Types.Struc;
    atol
) -> Tuple{Bool, Dict{Int64, Int64}}

Sym4state.Utils.magonly — Function

This function filters out the non-magnetic atoms in the struc structure. The atomic numbers of magnetic elements are provided in the mag_num_vec.

magonly(
    struc::Sym4state.Types.Struc,
    mag_num_vec::Vector{Int64}
) -> Sym4state.Types.Struc

Symmetric operations

This section contains functions working with space group and symmetric operations.

Sym4state.Utils.check_z_rot_mat — Function

This function examines a matrix and determines whether it represents a rotational matrix around the z-axis.

check_z_rot_mat(x::AbstractArray{T, 2}; atol) -> Bool

Sym4state.Utils.check_z_rot — Function

This function examines a SymOp and determines whether it contains a rotational matrix around the z-axis.

check_z_rot(op::Sym4state.Types.SymOp; atol) -> Bool

Sym4state.Utils.get_sym_op_vec — Function

This function using pymatgen.symmetry.analyzer from Pymatgen to get all the symmetric operations.

get_sym_op_vec(py_sga) -> Vector{Sym4state.Types.SymOp}

Atom pairs

This section contains functions working with atom pairs.

Sym4state.Utils.consider_pair_vec_in_radius — Function

This function provides the pairs of atoms that are within a specified cutoff radius, cutoff_radius, and include atoms from the center_idx_vec.

consider_pair_vec_in_radius(
    struc::Sym4state.Types.Struc,
    center_idx_vec::Vector{Int64},
    cutoff_radius::Float64
) -> Vector{Vector{Int64}}

Sym4state.Utils.linear_idx_to_vec — Function

This function computes the index of the primitive cell that encompasses the atom at index idx.

linear_idx_to_vec(
    idx::Int64,
    supercell_size::Vector{Int64},
    num_pri_sites::Int64
) -> Vector{Int64}

Sym4state.Utils.get_corresponding_pair_vec — Function

This function calculates the corresponding pair to the pair_vec after a specific operation.

get_corresponding_pair_vec(
    corresponding_dict::Dict{Int64, Int64},
    pair_vec::AbstractVector{Int64}
) -> Vector{Int64}

Sym4state.Utils.equal_pair — Function

This function determines all the equivalent atom pairs for a Struc that encloses the atom at index center_idx_vec within a specified cutoff radius of cutoff_radius. The symmetric operations are provided in the sym_op_vec.

equal_pair(
    mag_struc::Sym4state.Types.Struc,
    center_idx_vec::Vector{Int64},
    cutoff_radius::Float64,
    sym_op_vec::Vector{Sym4state.Types.SymOp};
    atol
) -> Union{Tuple{Bool, Nothing, Nothing}, Tuple{Bool, DataStructures.DisjointSets{Vector{Int64}}, Dict{Vector{Int64}, Sym4state.Types.SymOp}}}

Sym4state.Utils.get_fixed_pair_vec — Function

This function converts the atom pairs, which are identified by the indices of atoms in a supercell, into pairs characterized by the indices of atoms in a primitive cell along with the index of the corresponding primitive cells.

get_fixed_pair_vec(
    struc::Sym4state.Types.Struc,
    mag_num_vec::Vector{Int64},
    supercell_size::Vector{Int64},
    pair_vec::Vector{Int64}
) -> Vector{Int64}

Inputs and outputs

Sym4state.Utils.to_vasp_inputs — Function

With the help of Pymatgen, this function reorganizes the map_vec containing Map objects and the paths of the template INCAR, POSCAR, POTCAR, and KPOINTS files into separate directories representing different magnetic configurations for subsequent DFT calculations.

Additional specific INCAR settings can be provided through the kwargs parameter.

to_vasp_inputs(
    map_vec::Vector{Sym4state.Types.Map};
    incar_path,
    poscar_path,
    potcar_path,
    kpoints_path,
    kwargs...
)

Sym4state.Utils.get_py_struc — Function

This function using Pymatgen to read the POSCAR file at filepath.

get_py_struc(filepath::String) -> PythonCall.Core.Py

Sym4state.Utils.grep_energy — Function

This function retrieves the energy from the OSZICAR file located at oszicar_path.

grep_energy(oszicar_path) -> Float64

Sym4state.Utils.get_coeff_mat — Function

This function computes the interaction matrix utilizing the provided Map and the energies of various configurations stored in energy_vec.

get_coeff_mat(
    map::Sym4state.Types.Map,
    energy_vec::Dict{Int8, Float64}
) -> Matrix{Float64}

Sym4state.Utils.get_one_interact_coeff_mat — Function

This function calculates the interaction matrix using the provided Map and the directory cal_dir that contains the output files of DFT calculations.

get_one_interact_coeff_mat(
    map::Sym4state.Types.Map,
    cal_dir::String
) -> Matrix{Float64}

Sym4state.Utils.get_all_interact_coeff_under_sym — Function

This function computes all the equivalent interaction matrices within a group identified by groupid, employing coeffmat as the basis and considering the different relations stored in relation_vec.

get_all_interact_coeff_under_sym(
    coeff_mat::Matrix{Float64},
    group_idx::Int64,
    relation_vec::Vector{Sym4state.Types.CoeffMatRef}
) -> Tuple{Any, Any}

Sym4state.Utils.get_pair_and_coeff — Function

This function calculates all the interaction matrices and their corresponding atom pairs using the DFT calculation outputs stored in cal_dir. The map_vec stores the relations among magnetic configurations used for calculating each interaction matrix, while the relation_vec stores the relations among different interaction matrices within a group.

get_pair_and_coeff(
    map_vec::Vector{Sym4state.Types.Map},
    relation_vec::Vector{Sym4state.Types.CoeffMatRef},
    cal_dir::String
) -> Tuple{Matrix{Int64}, Array{Float64, 3}}

Miscellaneous

Sym4state.Utils.set_rwigs — Function

This function retrieves the RWIGS values for the provided path poscar_path of the POSCAR file, as specified in the referenced paper.

set_rwigs(poscar_path::String) -> Vector{Float64}

Sym4state.Utils.check_unit_cell — Function

This function checks whether the cutoff radius cutoff is big enough for the unit cell. This function is designed to bypass the errors when using CellListMap.jl in the reference of this file.

check_unit_cell(
    unit_cell_matrix::AbstractMatrix{Float64},
    cutoff::Float64
) -> Bool